requirements.txt
flask
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
torch


